Separated the history table into its own component file
